객실상세정보, 객실별 최근리뷰3개

+++ backend/src/main/java/com/lostay/backend/review/dto/ReviewDTO.java
@@ -25,6 +25,6 @@ public class ReviewDTO {
 	
 	private String roomName;					// 객실명
 	private double reviewAvg;					// 별점 평균
-	private int reviewCount;					// 리뷰 전체 개수
+	private long reviewCount;					// 리뷰 전체 개수
 	private String userNickname;				// 리뷰 작성자
 }
